+++ src/main/java/christmas/model/event/GiftEvent.java
@@ -8,7 +8,7 @@ public class GiftEvent implements Event{
 
     @Override
     public DiscountResult calculateDiscountAmount(Order order) {
-        long discountAmount = 0;
+        int discountAmount = 0;
         LocalDate startTime = LocalDate.of(2023, 12, 1);
         LocalDate endTime = LocalDate.of(2023, 12, 31);
